// src/api/generated/models/llmLogOverviewVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LlmModelCallCountVO } from './llmModelCallCountVO';

export interface LlmLogOverviewVO {
  totalTokens?: string;
  inputTokens?: string;
  outputTokens?: string;
  inputTokenRatio?: number;
  outputTokenRatio?: number;
  totalCalls?: string;
  successCalls?: string;
  failedCalls?: string;
  successRate?: number;
  averageDurationMillis?: number;
  modelCallCounts?: LlmModelCallCountVO[];
}