// src/api/generated/models/llmLogVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface LlmLogVO {
  id?: string;
  modelName?: string;
  userId?: string;
  status?: string;
  inputTokens?: string;
  outputTokens?: string;
  totalTokens?: string;
  usageJson?: string;
  durationMillis?: string;
  errorMessage?: string;
  createdAt?: string;
}