// src/api/generated/models/pageResultLlmLogVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LlmLogVO } from './llmLogVO';

export interface PageResultLlmLogVO {
  list?: LlmLogVO[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}